+++ apps/web/src/routes/_app/cards.tsx
@@ -5,6 +5,7 @@ import { RouteErrorFallback } from "@/components/error-message";
 import { Footer } from "@/components/layout/footer";
 import { Skeleton } from "@/components/ui/skeleton";
 import { catalogQueryOptions } from "@/hooks/use-cards";
+import { useHideScrollbar } from "@/hooks/use-hide-scrollbar";
 import { PAGE_PADDING } from "@/lib/utils";
 
 export const Route = createFileRoute("/_app/cards")({
@@ -15,6 +16,7 @@ export const Route = createFileRoute("/_app/cards")({
 });
 
 function CardsPage() {
+  useHideScrollbar();
   return (
     <>
       <div className={`flex flex-1 flex-col ${PAGE_PADDING}`}>